runtime: scaffold typed CST arithmetic wrappers
