+++ docs/design/adze-document.md
@@ -310,7 +310,7 @@ API direction so implementation PRs can stay small and reviewable:
 1. contract first,
 2. minimal document alpha,
 3. Tree-sitter projection over the document,
-4. typed CST spike,
+4. typed CST spike and generated-style wrapper scaffold,
 5. typed AST provenance,
 6. GLR ambiguity summaries,
 7. schema-versioned CLI/WASM outputs.
