+++ runtime/tests/typed_cst_arithmetic_spike.rs
@@ -2,159 +2,21 @@
 
 #![cfg(all(test, feature = "pure-rust", feature = "ts-compat"))]
 
-use adze::{
-    adze_ir::RuleId,
-    document::{AdzeDocument, NodeId, SyntaxNode},
-    parser_v4::Parser as CoreParser,
-    ts_compat::Language,
-};
-use std::sync::Arc;
+#[path = "support/typed_cst_arithmetic.rs"]
+mod typed_cst_arithmetic;
 
-fn fielded_arithmetic_language() -> Arc<Language> {
-    let mut lang = (*adze_example::ts_langs::arithmetic()).clone();
-    lang.table.field_names = vec![
-        "left".to_string(),
-        "operator".to_string(),
-        "right".to_string(),
-    ];
-    lang.table.field_map.insert((RuleId(2), 0), 0);
-    lang.table.field_map.insert((RuleId(2), 1), 1);
-    lang.table.field_map.insert((RuleId(2), 2), 2);
-    Arc::new(lang)
-}
-
-fn parse_fielded_document(source: &str) -> AdzeDocument {
-    let lang = fielded_arithmetic_language();
-    let mut parser = CoreParser::new(lang.grammar.clone(), lang.table.clone(), lang.name.clone());
-    parser
-        .parse_document(source)
-        .expect("document parse should succeed")
-}
-
-#[derive(Clone, Copy, Debug)]
-struct SourceFile<'doc> {
-    document: &'doc AdzeDocument,
-    id: NodeId,
-}
-
-impl<'doc> SourceFile<'doc> {
-    fn cast(document: &'doc AdzeDocument, id: NodeId) -> Option<Self> {
-        node_has_kind(document, id, "source_file").then_some(Self { document, id })
-    }
-
-    fn expression(&self) -> Option<Expression<'doc>> {
-        Expression::cast(self.document, self.child(0)?.node_id())
-    }
-}
-
-impl<'doc> SyntaxNode<'doc> for SourceFile<'doc> {
-    fn document(&self) -> &'doc AdzeDocument {
-        self.document
-    }
-
-    fn node_id(&self) -> NodeId {
-        self.id
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-struct Expression<'doc> {
-    document: &'doc AdzeDocument,
-    id: NodeId,
-}
-
-impl<'doc> Expression<'doc> {
-    fn cast(document: &'doc AdzeDocument, id: NodeId) -> Option<Self> {
-        node_has_kind(document, id, "expression").then_some(Self { document, id })
-    }
-
-    fn left(&self) -> Option<Expression<'doc>> {
-        let edge = self.edge_by_field_name("left")?;
-        Expression::cast(self.document, edge.child_id())
-    }
-
-    fn operator(&self) -> Option<MinusToken<'doc>> {
-        let edge = self.edge_by_field_name("operator")?;
-        MinusToken::cast(self.document, edge.child_id())
-    }
-
-    fn right(&self) -> Option<Expression<'doc>> {
-        let edge = self.edge_by_field_name("right")?;
-        Expression::cast(self.document, edge.child_id())
-    }
-
-    fn number(&self) -> Option<NumberToken<'doc>> {
-        NumberToken::cast(self.document, self.child(0)?.node_id())
-    }
-}
-
-impl<'doc> SyntaxNode<'doc> for Expression<'doc> {
-    fn document(&self) -> &'doc AdzeDocument {
-        self.document
-    }
-
-    fn node_id(&self) -> NodeId {
-        self.id
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-struct MinusToken<'doc> {
-    document: &'doc AdzeDocument,
-    id: NodeId,
-}
-
-impl<'doc> MinusToken<'doc> {
-    fn cast(document: &'doc AdzeDocument, id: NodeId) -> Option<Self> {
-        node_has_kind(document, id, "-").then_some(Self { document, id })
-    }
-}
-
-impl<'doc> SyntaxNode<'doc> for MinusToken<'doc> {
-    fn document(&self) -> &'doc AdzeDocument {
-        self.document
-    }
-
-    fn node_id(&self) -> NodeId {
-        self.id
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-struct NumberToken<'doc> {
-    document: &'doc AdzeDocument,
-    id: NodeId,
-}
-
-impl<'doc> NumberToken<'doc> {
-    fn cast(document: &'doc AdzeDocument, id: NodeId) -> Option<Self> {
-        node_has_kind(document, id, "number").then_some(Self { document, id })
-    }
-}
-
-impl<'doc> SyntaxNode<'doc> for NumberToken<'doc> {
-    fn document(&self) -> &'doc AdzeDocument {
-        self.document
-    }
-
-    fn node_id(&self) -> NodeId {
-        self.id
-    }
-}
-
-fn node_has_kind(document: &AdzeDocument, id: NodeId, expected: &str) -> bool {
-    document.tree().node(id).and_then(|node| node.kind_name()) == Some(expected)
-}
+use adze::document::SyntaxNode;
+use typed_cst_arithmetic::{parse_fielded_document, syntax};
 
 #[test]
-fn typed_cst_wrappers_project_from_document_node_ids_and_edges() {
+fn typed_cst_scaffold_projects_from_document_node_ids_and_edges() {
     let source = "1-2";
     let document = parse_fielded_document(source);
     let tree = document.tree();
     let root = tree.root();
     let expression_node = root.child(0).expect("root should expose expression child");
 
-    let syntax = SourceFile::cast(&document, tree.root_id()).expect("root should cast");
+    let syntax = syntax::source_file(&document).expect("root should cast");
     let expression = syntax
         .expression()
         .expect("source file should expose expression");
@@ -210,8 +72,8 @@ fn typed_cst_wrappers_project_from_document_node_ids_and_edges() {
         Some("2")
     );
 
-    assert!(SourceFile::cast(&document, expression.node_id()).is_none());
-    assert!(MinusToken::cast(&document, left.node_id()).is_none());
+    assert!(syntax::SourceFile::cast(&document, expression.node_id()).is_none());
+    assert!(syntax::MinusToken::cast(&document, left.node_id()).is_none());
     assert!(expression.edge_by_field_name("missing").is_none());
 }
 
@@ -223,7 +85,7 @@ fn typed_cst_wrappers_treat_recovered_mismatches_as_fallible_casts() {
     assert!(!document.diagnostics().is_empty());
     assert!(root.has_error());
 
-    let maybe_syntax = SourceFile::cast(&document, root.node_id());
+    let maybe_syntax = syntax::source_file(&document);
     if let Some(syntax) = maybe_syntax {
         let syntax_node = syntax.node().expect("source-file handle should resolve");
         assert_eq!(syntax.is_error(), syntax_node.is_error());
